починка site_capture

+++ src/site_capture/management/commands/refresh_images.py
@@ -20,28 +20,24 @@ class Command(BaseCommand):
 
         for faucet in faucets:
             position += 1
-            self.stdout.write('processing %s' % faucet.title_en)
-            self.stdout.write('there is {0} from {1} faucet'.format(position, faucets.count()))
-            image_name = str(uuid.uuid4()) + '.png'
-            image_path = os.path.join(faucets_images_root, image_name)
+            if "#" in faucet.image.name:
+                self.stdout.write('processing %s' % faucet.title_en)
+                self.stdout.write('there is {0} from {1} faucet'.format(position, faucets.count()))
+                image_name = str(uuid.uuid4()) + '.png'
+                image_path = os.path.join(faucets_images_root, image_name)
 
-            command = base_command.format(site=faucet.real_href(), output_file=image_path)
+                command = base_command.format(site=faucet.real_href(), output_file=image_path)
+                process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE)
 
-            process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE)
-
-            # make sure phantomjs has time to download/process the page
-            # but if we get nothing after 30 sec, just move on
-            try:
-                output, errors = process.communicate(timeout=20)
-                faucet.image.name = 'faucets/%s' % image_name
-                faucet.save()
-            except Exception as e:
-                faucet.image.name = 'faucets/%s' % image_name
-                faucet.save()
-                print("\t\tException: %s" % e)
                 try:
-                    process.kill()
-                except Exception as oloex:
-                    print("\t\tException: %s" % oloex)
+                    output, errors = process.communicate(timeout=60)
+                    faucet.image.name = 'faucets/%s' % image_name
+                    faucet.save()
+                except Exception as e:
+                    print("\t\tException: %s" % e)
+                    try:
+                        process.kill()
+                    except Exception as oloex:
+                        print("\t\tException: %s" % oloex)
 
         self.stdout.write('Everything done!')
